ci(actions): disable quarto render (as not found fix for including .Rmd without env error)
This error has been explored in detail in my reproduction of Huang et al. 2019, but I've yet to find a resolution. The issue is that the renv and python environment build successfully but, when rendering the quarto site, it then cannot find the renv. My current solution is to disable this action and instead manually update the quarto site using quarto publish from local terminal.